Fix permission denied when live syncing to android device

+++ lib/services/usb-livesync-service.ts
@@ -107,7 +107,9 @@ export class AndroidUsbLiveSyncService extends androidLiveSyncServiceLib.Android
 				let devicePathRoot = `/data/data/${deviceAppData.appIdentifier}/files`;
 				_.each(localToDevicePaths, localToDevicePath => {
 					let devicePath = this.$mobileHelper.correctDevicePath(path.join(devicePathRoot, localToDevicePath.getRelativeToProjectBasePath()));
-					commands.push(`mv "${localToDevicePath.getDevicePath()}" "${devicePath}"`);
+					if(this.$fs.getFsStats(localToDevicePath.getLocalPath()).wait().isFile()) {
+						commands.push(`mv "${localToDevicePath.getDevicePath()}" "${devicePath}"`);
+					}
 				});
 				
 				commands.push(`rm -rf ${this.$mobileHelper.buildDevicePath(devicePathRoot, "code_cache", "secondary_dexes", "proxyThumb")}`);
